+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 # Change history for ui-licenses
 
-## 2.0.0 IN PROGRESS
+## 2.2.0
+* Added name export of Licenses search UI
+
+## 2.1.0
+* Added ability to view, add, edit, and remove core documents from licenses.
+
+## 2.0.0
 * Upgrade to Stripes 2.0
 * Removed usage of lookbehinds in regexes (avail in 2.0.1)
 * Added License Status filter to search.
 * Added ability to view, add, edit, and remove terms from licenses.
 * Added ability to view, add, edit, and remove organizations from licenses.
 * Added sticky license header.
-* Added ability to view, add, edit, and remove core documents from licenses.
 
 ## 1.1.0
 First Official Release for Q4 Release - 2018-12-11